Add files via upload
This small Python project was designed to find the friend-communication-groups for Poets in Tang Dynasty of China under the Harvard CBDB Work Group.
 
Three classical algorithms are used in this piece of code: Binary-search, minimum spanning trees, and depth-first-search. 
Some code segments are obtained from public source with little modification. 

The code is posted for public academic use as it is. The programmer/or Harvard CBDB takes no responsibility for any concerns regarding to the code.
 
Contact with the Harvard CBDB Work Group for database format issue. 
The code was written by Dr. Li Chen (lchen@udc.edu) at the University of the District of Columbia.
 
 
Version 1, revision 1.
 7/25/2021